moves countiesArr and statesArr out of CompareTable. re-orders imports
