+++ src/components/LocationPage/ChartsHolder.tsx
@@ -10,6 +10,8 @@ import useMediaQuery from '@material-ui/core/useMediaQuery';
 import { useTheme } from '@material-ui/core/styles';
 import { Metric } from 'common/metric';
 import CompareMain from 'components/Compare/CompareMain';
+import { getCountiesArr } from 'common/utils/compare';
+import { countySummary } from 'common/location_summaries';
 
 // TODO(michael): figure out where this type declaration should live.
 type County = {
@@ -147,6 +149,12 @@ const ChartsHolder = (props: {
       ]
     : [];
 
+  const locationsForCompare = getCountiesArr(props.stateId);
+  const currentCountyForCompare: any = props.county && {
+    locationInfo: props.county,
+    metricsInfo: countySummary(props.county.full_fips_code),
+  };
+
   // TODO(pablo): Create separate refs for signup and share
   return (
     <>
@@ -176,6 +184,8 @@ const ChartsHolder = (props: {
               county={props.county}
               locationsViewable={10}
               isLocationPage
+              locations={locationsForCompare}
+              currentCounty={currentCountyForCompare}
             />
             <MainContentInner>
               {chartPropsForMap.map(chartProps => (
